+++ server/src/main/java/com/enegoce/repository/DealGoodsRepository.java
@@ -9,9 +9,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DealGoodsRepository extends JpaRepository<DealGoods, Integer> {
+public interface DealGoodsRepository extends JpaRepository<DealGoods, Long> {
 
     @Query("SELECT dg FROM DealGoods dg WHERE dg.id.deal.id = :dealId")
-    List<DealGoods> findGoodsByDealId(@Param("dealId") Integer dealId);
+    List<DealGoods> findGoodsByDealId(@Param("dealId") Long dealId);
 
 }
