+++ server/src/main/java/com/enegoce/repository/DealPartyRepository.java
@@ -9,11 +9,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 @Repository
-public interface DealPartyRepository extends JpaRepository<DealParty, Integer> {
+public interface DealPartyRepository extends JpaRepository<DealParty, Long> {
 
     @Query("SELECT dp FROM DealParty dp WHERE dp.id.deal.id = :dealId")
-    List<DealParty> findPartiesByDealId(@Param("dealId") Integer dealId);
+    List<DealParty> findPartiesByDealId(@Param("dealId") Long dealId);
 
     @Query("SELECT dp FROM DealParty dp WHERE dp.id.deal.id = :dealId AND dp.id.codPrt LIKE :code")
-    DealParty findPartyByDealIdAndCode(@Param("dealId") Integer dealId, @Param("code") String code);
+    DealParty findPartyByDealIdAndCode(@Param("dealId") Long dealId, @Param("code") String code);
 }
